services/report-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0

